+++ web-frontend/modules/core/plugins/clientHandler.js
@@ -127,6 +127,10 @@ export class ClientErrorMap {
         app.i18n.t('clientHandler.cannotDisableAllAuthProvidersTitle'),
         app.i18n.t('clientHandler.cannotDisableAllAuthProvidersDescription')
       ),
+      ERROR_LAST_ADMIN_OF_GROUP: new ResponseErrorMessage(
+        app.i18n.t('clientHandler.lastAdminTitle'),
+        app.i18n.t('clientHandler.lastAdminMessage')
+      ),
     }
   }